+++ src/components/common/checkbox/checkboxGroup.tsx
@@ -1,4 +1,4 @@
-import { FC } from "react";
+import { FC, useCallback, useState } from "react";
 import { useField, UseFieldHookProps } from "packages/form";
 import FormLabel, { FormLabelProps } from "@material-ui/core/FormLabel";
 import FormGroup, { FormGroupProps } from "@material-ui/core/FormGroup";
@@ -12,10 +12,15 @@ import Grid, { GridProps } from "@material-ui/core/Grid";
 import FormHelperText, {
   FormHelperTextProps,
 } from "@material-ui/core/FormHelperText";
-import { Merge, OptionsProps } from "../types";
+import CircularProgress, {
+  CircularProgressProps,
+} from "@material-ui/core/CircularProgress";
+import InputAdornment from "@material-ui/core/InputAdornment";
+import { Merge, OptionsProps, dependentOptionsFn } from "../types";
+import { getLabelFromValues, useOptionsFetcher } from "../utils";
 
 interface extendedFiledProps extends UseFieldHookProps {
-  options: OptionsProps[];
+  options: OptionsProps[] | dependentOptionsFn;
   label: string;
 }
 
@@ -29,6 +34,7 @@ interface MyCheckboxExtendedProps {
   FormHelperTextProps?: FormHelperTextProps;
   GridProps?: GridProps;
   enableGrid: boolean;
+  CircularProgressProps?: CircularProgressProps;
 }
 
 export type MyCheckboxGroupAllProps = Merge<
@@ -60,9 +66,11 @@ const MyCheckboxGroup: FC<MyCheckboxGroupAllProps> = ({
   GridProps,
   enableGrid,
   runValidationOnDependentFieldsChange,
+  CircularProgressProps,
   ...others
 }) => {
   const {
+    formState,
     value,
     error,
     touched,
@@ -73,6 +81,10 @@ const MyCheckboxGroup: FC<MyCheckboxGroupAllProps> = ({
     name,
     excluded,
     readOnly,
+    dependentValues,
+    incomingMessage,
+    runValidation,
+    whenToRunValidation,
   } = useField({
     name: fieldName,
     fieldKey: fieldID,
@@ -85,11 +97,37 @@ const MyCheckboxGroup: FC<MyCheckboxGroupAllProps> = ({
     shouldExclude,
     runValidationOnDependentFieldsChange,
   });
+  const [_options, setOptions] = useState<OptionsProps[]>([]);
+
+  const getLabelFromValuesForOptions = useCallback(
+    (values) => getLabelFromValues(_options)(values),
+    [_options]
+  );
+
+  const handleChangeInterceptor = useCallback(
+    (e) => {
+      const value = typeof e === "object" ? e?.target?.value ?? "" : e;
+      let result = getLabelFromValuesForOptions(value);
+      handleChange(e, result[0] as any);
+    },
+    [handleChange, getLabelFromValuesForOptions]
+  );
+  const { loadingOptions } = useOptionsFetcher(
+    formState,
+    options,
+    setOptions,
+    handleChangeInterceptor,
+    dependentValues,
+    incomingMessage,
+    runValidation,
+    whenToRunValidation
+  );
+
   if (excluded) {
     return null;
   }
   const isError = touched && (error ?? "") !== "";
-  const checkboxes = options.map((checkbox) => (
+  const checkboxes = _options.map((checkbox) => (
     <FormControlLabel
       {...FormControlLabelProps}
       control={
@@ -101,7 +139,7 @@ const MyCheckboxGroup: FC<MyCheckboxGroupAllProps> = ({
       }
       key={checkbox.value}
       name={name}
-      onChange={handleChange}
+      onChange={handleChangeInterceptor}
       label={checkbox.label}
       value={checkbox.value}
       checked={valueExists(value, checkbox.value)}
@@ -120,7 +158,19 @@ const MyCheckboxGroup: FC<MyCheckboxGroupAllProps> = ({
       <FormLabel {...FormLabelProps} component="label">
         {label}
       </FormLabel>
-      <FormGroup {...FormGroupProps}>{checkboxes}</FormGroup>
+      <FormGroup {...FormGroupProps}>
+        {loadingOptions ? (
+          <InputAdornment position="end">
+            <CircularProgress
+              color="primary"
+              variant="indeterminate"
+              {...CircularProgressProps}
+            />
+          </InputAdornment>
+        ) : (
+          checkboxes
+        )}
+      </FormGroup>
       {isError ? (
         <FormHelperText {...FormHelperTextProps}>{error}</FormHelperText>
       ) : null}
