+++ src/components/common/radio/radio.tsx
@@ -1,4 +1,4 @@
-import { FC } from "react";
+import { FC, useCallback, useState } from "react";
 import { useField, UseFieldHookProps } from "packages/form";
 import FormControl, { FormControlProps } from "@material-ui/core/FormControl";
 import FormLabel, { FormLabelProps } from "@material-ui/core/FormLabel";
@@ -12,7 +12,12 @@ import FormHelperText, {
   FormHelperTextProps,
 } from "@material-ui/core/FormHelperText";
 import Grid, { GridProps } from "@material-ui/core/Grid";
+import CircularProgress, {
+  CircularProgressProps,
+} from "@material-ui/core/CircularProgress";
+import InputAdornment from "@material-ui/core/InputAdornment";
 import { Merge, OptionsProps } from "../types";
+import { getLabelFromValues, useOptionsFetcher } from "../utils";
 
 interface extendedFiledProps extends UseFieldHookProps {
   options: OptionsProps[];
@@ -29,6 +34,7 @@ interface MyCheckboxExtendedProps {
   FormHelperTextProps?: FormHelperTextProps;
   GridProps?: GridProps;
   enableGrid: boolean;
+  CircularProgressProps?: CircularProgressProps;
 }
 
 export type MyRadioAllProps = Merge<MyRadioMixedProps, MyCheckboxExtendedProps>;
@@ -53,9 +59,11 @@ const MyRadio: FC<MyRadioAllProps> = ({
   GridProps,
   enableGrid,
   runValidationOnDependentFieldsChange,
+  CircularProgressProps,
   ...others
 }) => {
   const {
+    formState,
     value,
     error,
     touched,
@@ -66,6 +74,10 @@ const MyRadio: FC<MyRadioAllProps> = ({
     name,
     excluded,
     readOnly,
+    dependentValues,
+    incomingMessage,
+    runValidation,
+    whenToRunValidation,
   } = useField({
     name: fieldName,
     fieldKey: fieldID,
@@ -78,11 +90,37 @@ const MyRadio: FC<MyRadioAllProps> = ({
     shouldExclude,
     runValidationOnDependentFieldsChange,
   });
+  const [_options, setOptions] = useState<OptionsProps[]>([]);
+
+  const getLabelFromValuesForOptions = useCallback(
+    (values) => getLabelFromValues(_options)(values),
+    [_options]
+  );
+
+  const handleChangeInterceptor = useCallback(
+    (e) => {
+      const value = typeof e === "object" ? e?.target?.value ?? "" : e;
+      let result = getLabelFromValuesForOptions(value);
+      handleChange(e, result[0] as any);
+    },
+    [handleChange, getLabelFromValuesForOptions]
+  );
+  const { loadingOptions } = useOptionsFetcher(
+    formState,
+    options,
+    setOptions,
+    handleChangeInterceptor,
+    dependentValues,
+    incomingMessage,
+    runValidation,
+    whenToRunValidation
+  );
+
   if (excluded) {
     return null;
   }
   const isError = touched && (error ?? "") !== "";
-  const radios = options.map((radio) => (
+  const radios = _options.map((radio) => (
     <FormControlLabel
       {...FormControlLabelProps}
       control={
@@ -112,11 +150,21 @@ const MyRadio: FC<MyRadioAllProps> = ({
       </FormLabel>
       <RadioGroup
         {...RadioGroupProps}
-        onChange={handleChange}
+        onChange={handleChangeInterceptor}
         name={name}
         value={value}
       >
-        {radios}
+        {loadingOptions ? (
+          <InputAdornment position="end">
+            <CircularProgress
+              color="primary"
+              variant="indeterminate"
+              {...CircularProgressProps}
+            />
+          </InputAdornment>
+        ) : (
+          radios
+        )}
       </RadioGroup>
       {isError ? (
         <FormHelperText {...FormHelperTextProps}>{error}</FormHelperText>
