+++ src/components/common/utils/optionsFetcher.ts
@@ -1,0 +1,80 @@
+import { useState, useRef, useCallback, useEffect } from "react";
+
+export const useOptionsFetcher = (
+  formState,
+  options,
+  setOptions,
+  handleChangeInterceptor,
+  dependentValues,
+  incomingMessage,
+  runValidation,
+  whenToRunValidation
+) => {
+  const lastOptionsPromise = useRef<Promise<any> | null>(null);
+  const [loadingOptions, setLoadingOptions] = useState(false);
+
+  const syncAsyncSetOptions = useCallback(
+    (options, dependentValues) => {
+      if (Array.isArray(options)) {
+        setOptions(options);
+      } else if (typeof options === "function") {
+        try {
+          setLoadingOptions(true);
+          setOptions([{ label: "loading...", value: null }]);
+          let currentPromise = Promise.resolve(
+            options(dependentValues, formState)
+          );
+          lastOptionsPromise.current = currentPromise;
+          currentPromise
+            .then((result) => {
+              setLoadingOptions(false);
+              if (lastOptionsPromise.current === currentPromise) {
+                if (Array.isArray(result)) {
+                  setOptions(result);
+                } else {
+                  setOptions([{ label: "Couldn't fetch", value: null }]);
+                  console.log(
+                    `expected optionsFunction in select component to return array of OptionsType but got: ${result}`
+                  );
+                }
+              }
+            })
+            .catch((e) => {
+              setLoadingOptions(false);
+              setOptions([{ label: "Couldn't fetch", value: null }]);
+              console.log(`error occured while fetching options`, e?.message);
+            });
+        } catch (e) {
+          setLoadingOptions(false);
+          setOptions([{ label: "Couldn't fetch", value: null }]);
+          console.log(`error occured while fetching options`, e?.message);
+        }
+      }
+    },
+    [setOptions, formState]
+  );
+  useEffect(() => {
+    syncAsyncSetOptions(options, dependentValues);
+  }, [options, dependentValues, syncAsyncSetOptions]);
+
+  useEffect(() => {
+    if (incomingMessage !== null && typeof incomingMessage === "object") {
+      const { value, options } = incomingMessage;
+      handleChangeInterceptor(value);
+      if (whenToRunValidation === "onBlur") {
+        runValidation({ value: value }, true);
+      }
+      if (Array.isArray(options)) {
+        setOptions(options);
+      }
+    }
+  }, [
+    incomingMessage,
+    setOptions,
+    handleChangeInterceptor,
+    runValidation,
+    whenToRunValidation,
+  ]);
+
+  return { loadingOptions };
+};
